Move each state to a new module
