+++ apps/src/lib/node/ledger/shell/prepare_proposal/block_space_alloc.rs
@@ -101,211 +101,6 @@ impl BlockSpaceAllocator<states::BuildingDecryptedTxBatch> {
     }
 }
 
-impl states::State for BlockSpaceAllocator<states::BuildingDecryptedTxBatch> {
-    type Next = BlockSpaceAllocator<states::BuildingProtocolTxBatch>;
-
-    #[inline]
-    fn try_alloc(&mut self, tx: &[u8]) -> AllocStatus {
-        self.decrypted_txs.try_dump(tx)
-    }
-
-    #[inline]
-    fn try_alloc_batch<'tx, T>(&mut self, txs: T) -> AllocStatus
-    where
-        T: IntoIterator<Item = &'tx [u8]> + 'tx,
-    {
-        self.decrypted_txs.try_dump_all(txs)
-    }
-
-    #[inline]
-    fn next_state(mut self) -> Self::Next {
-        // seal decrypted txs
-        self.decrypted_txs.allotted_space_in_bytes =
-            self.decrypted_txs.current_space_in_bytes;
-
-        // reserve space for protocol txs
-        let uninit = self.uninitialized_space_in_bytes();
-        self.protocol_txs = TxBin::init_over_ratio(uninit, thres::ONE_THIRD);
-
-        // cast state
-        let Self {
-            max_block_space_in_bytes,
-            protocol_txs,
-            encrypted_txs,
-            decrypted_txs,
-            ..
-        } = self;
-
-        BlockSpaceAllocator {
-            _state: PhantomData,
-            max_block_space_in_bytes,
-            protocol_txs,
-            encrypted_txs,
-            decrypted_txs,
-        }
-    }
-}
-
-impl states::State<states::WithEncryptedTxs>
-    for BlockSpaceAllocator<states::BuildingProtocolTxBatch>
-{
-    type Next = BlockSpaceAllocator<
-        states::BuildingEncryptedTxBatch<states::WithEncryptedTxs>,
-    >;
-
-    #[inline]
-    fn try_alloc(&mut self, tx: &[u8]) -> AllocStatus {
-        self.protocol_txs.try_dump(tx)
-    }
-
-    #[inline]
-    fn try_alloc_batch<'tx, T>(&mut self, txs: T) -> AllocStatus
-    where
-        T: IntoIterator<Item = &'tx [u8]> + 'tx,
-    {
-        self.protocol_txs.try_dump_all(txs)
-    }
-
-    #[inline]
-    fn next_state(self) -> Self::Next {
-        todo!()
-    }
-}
-
-impl states::State<states::WithoutEncryptedTxs>
-    for BlockSpaceAllocator<states::BuildingProtocolTxBatch>
-{
-    type Next = BlockSpaceAllocator<
-        states::BuildingEncryptedTxBatch<states::WithoutEncryptedTxs>,
-    >;
-
-    #[inline]
-    fn try_alloc(&mut self, tx: &[u8]) -> AllocStatus {
-        self.protocol_txs.try_dump(tx)
-    }
-
-    #[inline]
-    fn try_alloc_batch<'tx, T>(&mut self, txs: T) -> AllocStatus
-    where
-        T: IntoIterator<Item = &'tx [u8]> + 'tx,
-    {
-        self.protocol_txs.try_dump_all(txs)
-    }
-
-    #[inline]
-    fn next_state(self) -> Self::Next {
-        todo!()
-    }
-}
-
-impl states::State
-    for BlockSpaceAllocator<
-        states::BuildingEncryptedTxBatch<states::WithEncryptedTxs>,
-    >
-{
-    type Next = BlockSpaceAllocator<
-        states::FillingRemainingSpace<states::WithEncryptedTxs>,
-    >;
-
-    #[inline]
-    fn try_alloc(&mut self, tx: &[u8]) -> AllocStatus {
-        self.encrypted_txs.try_dump(tx)
-    }
-
-    #[inline]
-    fn try_alloc_batch<'tx, T>(&mut self, txs: T) -> AllocStatus
-    where
-        T: IntoIterator<Item = &'tx [u8]> + 'tx,
-    {
-        self.encrypted_txs.try_dump_all(txs)
-    }
-
-    #[inline]
-    fn next_state(self) -> Self::Next {
-        todo!()
-    }
-}
-
-impl states::State
-    for BlockSpaceAllocator<
-        states::BuildingEncryptedTxBatch<states::WithoutEncryptedTxs>,
-    >
-{
-    type Next = BlockSpaceAllocator<
-        states::FillingRemainingSpace<states::WithoutEncryptedTxs>,
-    >;
-
-    #[inline]
-    fn try_alloc(&mut self, tx: &[u8]) -> AllocStatus {
-        self.encrypted_txs.try_dump(tx)
-    }
-
-    #[inline]
-    fn try_alloc_batch<'tx, T>(&mut self, txs: T) -> AllocStatus
-    where
-        T: IntoIterator<Item = &'tx [u8]> + 'tx,
-    {
-        self.encrypted_txs.try_dump_all(txs)
-    }
-
-    #[inline]
-    fn next_state(self) -> Self::Next {
-        todo!()
-    }
-}
-
-impl states::State
-    for BlockSpaceAllocator<
-        states::FillingRemainingSpace<states::WithEncryptedTxs>,
-    >
-{
-    type Next = ();
-
-    #[inline]
-    fn try_alloc(&mut self, _tx: &[u8]) -> AllocStatus {
-        todo!()
-    }
-
-    #[inline]
-    fn try_alloc_batch<'tx, T>(&mut self, _txs: T) -> AllocStatus
-    where
-        T: IntoIterator<Item = &'tx [u8]> + 'tx,
-    {
-        todo!()
-    }
-
-    #[inline]
-    fn next_state(self) -> Self::Next {
-        // NOOP
-    }
-}
-
-impl states::State
-    for BlockSpaceAllocator<
-        states::FillingRemainingSpace<states::WithoutEncryptedTxs>,
-    >
-{
-    type Next = ();
-
-    #[inline]
-    fn try_alloc(&mut self, _tx: &[u8]) -> AllocStatus {
-        todo!()
-    }
-
-    #[inline]
-    fn try_alloc_batch<'tx, T>(&mut self, _txs: T) -> AllocStatus
-    where
-        T: IntoIterator<Item = &'tx [u8]> + 'tx,
-    {
-        todo!()
-    }
-
-    #[inline]
-    fn next_state(self) -> Self::Next {
-        // NOOP
-    }
-}
-
 impl<State> BlockSpaceAllocator<State> {
     /// Return uninitialized space in tx bins, resulting from ratio conversions.
     ///
