+++ apps/src/lib/node/ledger/shell/prepare_proposal/block_space_alloc/states/encrypted_txs.rs
@@ -1,0 +1,51 @@
+use super::super::{AllocStatus, BlockSpaceAllocator};
+use super::{
+    BuildingEncryptedTxBatch, FillingRemainingSpace, State, WithEncryptedTxs,
+    WithoutEncryptedTxs,
+};
+
+impl State for BlockSpaceAllocator<BuildingEncryptedTxBatch<WithEncryptedTxs>> {
+    type Next = BlockSpaceAllocator<FillingRemainingSpace<WithEncryptedTxs>>;
+
+    #[inline]
+    fn try_alloc(&mut self, tx: &[u8]) -> AllocStatus {
+        self.encrypted_txs.try_dump(tx)
+    }
+
+    #[inline]
+    fn try_alloc_batch<'tx, T>(&mut self, txs: T) -> AllocStatus
+    where
+        T: IntoIterator<Item = &'tx [u8]> + 'tx,
+    {
+        self.encrypted_txs.try_dump_all(txs)
+    }
+
+    #[inline]
+    fn next_state(self) -> Self::Next {
+        todo!()
+    }
+}
+
+impl State
+    for BlockSpaceAllocator<BuildingEncryptedTxBatch<WithoutEncryptedTxs>>
+{
+    type Next = BlockSpaceAllocator<FillingRemainingSpace<WithoutEncryptedTxs>>;
+
+    #[inline]
+    fn try_alloc(&mut self, tx: &[u8]) -> AllocStatus {
+        self.encrypted_txs.try_dump(tx)
+    }
+
+    #[inline]
+    fn try_alloc_batch<'tx, T>(&mut self, txs: T) -> AllocStatus
+    where
+        T: IntoIterator<Item = &'tx [u8]> + 'tx,
+    {
+        self.encrypted_txs.try_dump_all(txs)
+    }
+
+    #[inline]
+    fn next_state(self) -> Self::Next {
+        todo!()
+    }
+}
